old/requirements.txt
cloudmesh_client
cloudmesh_workflow
jinja2schema
django-rest-swagger
django-bootstrap3
django-admin-bootstrapped
django-bootstrap-themes
django-jinja
djangorestframework
markdown
django-filter
aldjemy
nwdiag
pygal
django-jinja-bootstrap-form
